+++ src/lte/model/lte-enb-cphy-sap.h
@@ -63,7 +63,7 @@ public:
    * \param ulEarfcn the UL EARFCN
    * \param dlEarfcn the DL EARFCN
    */
-  virtual void SetEarfcn (uint16_t ulEarfcn, uint16_t dlEarfcn) = 0;
+  virtual void SetEarfcn (uint32_t ulEarfcn, uint32_t dlEarfcn) = 0;
   
   /** 
    * Add a new UE to the cell
@@ -151,7 +151,7 @@ public:
   // inherited from LteEnbCphySapProvider
   virtual void SetCellId (uint16_t cellId);
   virtual void SetBandwidth (uint8_t ulBandwidth, uint8_t dlBandwidth);
-  virtual void SetEarfcn (uint16_t ulEarfcn, uint16_t dlEarfcn);
+  virtual void SetEarfcn (uint32_t ulEarfcn, uint32_t dlEarfcn);
   virtual void AddUe (uint16_t rnti);
   virtual void RemoveUe (uint16_t rnti);
   virtual void SetPa (uint16_t rnti, double pa);
@@ -194,7 +194,7 @@ MemberLteEnbCphySapProvider<C>::SetBandwidth (uint8_t ulBandwidth, uint8_t dlBan
 
 template <class C>
 void 
-MemberLteEnbCphySapProvider<C>::SetEarfcn (uint16_t ulEarfcn, uint16_t dlEarfcn)
+MemberLteEnbCphySapProvider<C>::SetEarfcn (uint32_t ulEarfcn, uint32_t dlEarfcn)
 {
   m_owner->DoSetEarfcn (ulEarfcn, dlEarfcn);
 }
